apple-gmux: Resume discrete GPU on hotplug
gmux receives the HPD pin of external DP and HDMI ports as well as DP
sink ports of built-in Thunderbolt controllers. (On retinas, where the
panel is connected with eDP, it also receives the panel's HPD pin.)
gmux raises an interrupt when hotplug is detected on one of those pins.

On MacBook Pros introduced before 2011, the external DP port is fully
switchable between GPUs and we always switch it in unison with the
panel. The GPU to which the external DP port is switched is therefore
guaranteed to be powered up and can detect hotplug by itself.

On MacBook Pros introduced 2011 and onward, external DP ports are
combined DP/Thunderbolt ports that are no longer fully switchable
between GPUs, they can only be driven by the discrete GPU. If the panel
is switched to the discrete GPU, the GPU is powered up and can detect
hotplug by itself. However if the panel is switched to the integrated
GPU, the discrete GPU is usually powered down and therefore unable to
detect hotplug.

Resume the discrete GPU in this situation.

Signed-off-by: Lukas Wunner <lukas@wunner.de>