thunderbolt: Power down controller on Macs when nothing is plugged in
Document and implement Apple's ACPI-based (but nonstandard) PM mechanism
for Thunderbolt 1 and 2.  Briefly, an ACPI method provided by Apple is
used to cut power to the controller.  A GPE is enabled while the
controller is powered down which sideband-signals a plug event,
whereupon power is reinstated using the ACPI method.

This saves 1.5 W on machines with a Light Ridge controller and is
reported to save 4 W on Cactus Ridge 4C and Falcon Ridge 4C.  (I believe
4 W includes the bus power drawn by Apple's Gigabit Ethernet adapter.)
It fixes a power regression introduced in 3.17 by commit 7bc5a2bad0b8
("ACPI: Support _OSI("Darwin") correctly").

A Thunderbolt controller appears to the OS as a set of PCI devices:
One upstream bridge, multiple downstream bridges and one NHI (Native
Host Interface).  The upstream and downstream bridges represent a PCIe
switch (see definition of a switch in the PCIe spec).  The NHI device is
used to manage the Converged I/O switch over which PCI tunnels are set
up to hotplugged devices.  Those devices in turn appear behind the
downstream bridges:

  (Root Port) ---- Upstream Bridge --+-- Downstream Bridge 0 ---- NHI
                                     +-- Downstream Bridge 1 --
                                     +-- Downstream Bridge 2 --
                                     ...

Power is cut to the entire set of devices.  The Linux pm model is
hierarchical and assumes that a child cannot resume before its parent.
To conform to this model, power control must be governed by the
Thunderbolt controller's topmost device, which is the upstream bridge.
The NHI and downstream bridges go to D3hot independently and the
upstream bridge goes to D3cold once all its children have suspended.
Hence this commit modifies runtime PM of the upstream bridge.

Because Apple's ACPI methods are nonstandard, a struct dev_pm_domain is
used to override the PCI bus pm_ops.  The thunderbolt driver binds to
the NHI, thus the dev_pm_domain is assigned to the upstream bridge when
its grandchild ->probes and is evicted when it ->removes.

The NHI is not able to detect plug events while suspended, it relies on
the GPE handler to resume it on hotplug.  A runtime PM ref is held for
the duration of tb_handle_hotplug() to keep the NHI awake as long as a
hotplug event is processed.  Apart from that a runtime PM ref is held
for each attached Thunderbolt switch to ensure the NHI stays runtime
active as long as devices are plugged in.  This behaviour is identical
to the macOS driver.  It implies that RTD3 (runtime D3cold with devices
attached) is not available on Macs.

Handles for the ACPI methods and various other data is stored in a newly
introduced struct tb_pm which is pointed to by struct tb.  This data is
not specific to OS-controlled tunnel management:  Thunderbolt 3 Macs
(which use ICM-controlled tunnel management) use a similar PM mechanism.
If the need arises to store additional or different PM data for Macs or
non-Macs with Thunderbolt 3 efficiently then struct tb_pm can be changed
to a union.

There are no Thunderbolt specs publicly available from Intel or Apple,
so I've included documentation to the extent that I was able to reverse-
engineer things.  Documentation on the Go2Sx and Ok2Go2Sx pins is
tentative as those are missing on my Light Ridge.  Apple only uses them
on Cactus Ridge 4C.  Someone with such a controller needs to find out
through experimentation if the documentation is accurate and amend it if
necessary.

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=92111
Signed-off-by: Lukas Wunner <lukas@wunner.de>