thunderbolt: Runtime suspend NHI when idle
Runtime suspend the NHI when no Thunderbolt devices have been plugged in
for 10 sec (user-configurable via autosuspend_delay_ms in sysfs).

The NHI is not able to detect plug events while suspended, it relies on
the GPE handler to resume it on hotplug.

After the NHI resumes, it takes about 700 ms until a hotplug event
appears on the RX ring.  In case autosuspend_delay_ms has been reduced
to 0 by the user, we need to wait in tb_resume() to avoid going back to
sleep before we had a chance to detect a hotplugged device.  A runtime
pm ref is held for the duration of tb_handle_hotplug() to keep the NHI
awake while the hotplug event is processed.

Apart from that we acquire a runtime pm ref for each newly allocated
switch (except for the root switch) and drop one when a switch is freed,
thereby ensuring the NHI stays active as long as devices are plugged in.
This behaviour is identical to the macOS driver.

One open issue remains:

* On Cactus Ridge and newer, directly-attached DisplayPort devices are
  not wired to the GPU, but rather the DP signal is routed through the
  Thunderbolt controller.  It may thus be necessary to check presence of
  attached DisplayPort devices when scanning the root switch and
  acquire a runtime ref for each of them to prevent the controller
  from powering off while displays are attached.  Unfortunately my own
  machine predates Cactus Ridge and wires the DP signal directly to the
  GPU, so I cannot test this.

Cc: Andreas Noever <andreas.noever@gmail.com>
Cc: Michael Jamet <michael.jamet@intel.com>
Cc: Mika Westerberg <mika.westerberg@linux.intel.com>
Cc: Yehezkel Bernat <yehezkel.bernat@intel.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>