thunderbolt: Use device links instead of PCI quirk
When resuming from system sleep, the PCI hierarchy is walked down during
the ->resume_noirq phase.  Those subtrees of the hierarchy which are
behind a Thunderbolt PCI tunnel are inaccessible until the tunnel has
been re-established by the NHI driver.  Hence a Thunderbolt controller's
hotplug bridges (below which the devices behind the tunnel appear) must
delay resuming until the NHI has finished.

That requirement is not enforced by the PM core automatically as it only
guarantees resume ordering between parent and child, and the NHI is not
a parent of the hotplug bridges, but a niece.

So far we've open coded this requirement in a PCI quirk, which has the
following disadvantages:

- The code for the NHI resides in drivers/thunderbolt/, whereas the code
  for the hotplug bridges resides in drivers/pci/quirks.c, which may be
  surprising and non-obvious in particular for new contributors.

- Whenever support for an additional Thunderbolt controller is added,
  its PCI device ID needs to be amended in both places, which invites
  mistakes.  E.g. commit a42fb351ca1f ("thunderbolt: Allow loading of
  module on recent Apple MacBooks with thunderbolt 2 controller")
  relaxed the subvendor and subdevice ID in the NHI code but forgot to
  also change the hotplug bridge code.

- Since the PCI quirk cannot keep any state, it has to search for the
  NHI over and over again on every resume.

- Thunderbolt PCI device IDs need to be kept in include/linux/pci_ids.h
  rather than being moved to a header in drivers/thunderbolt/ because
  they're used both by the NHI code and hotplug bridge code.

Starting with v4.10, the PM core has gained the ability to declare
"device links", i.e. dependencies between devices beyond the mere
parent/child relationship.

Replace the PCI quirk with device links from the hotplug bridges to the
NHI.  The device links are set up when loading the thunderbolt driver
and torn down when it is removed.  The PM core takes care of everything
else.  As a bonus, the amount of code is cut in half.

The device links are also taken into account for runtime PM, which
solves corner cases where the NHI suspends but the hotplug bridges do
not (e.g. because pciehp_poll_mode is enabled):  The NHI isn't woken on
hotplug unless the wake GPE is enabled, the wake GPE can only be enabled
after the chip has been powered down, and powering down the chip is
prevented by the hotplug bridges in poll mode.

Cc: Rafael J. Wysocki <rjw@rjwysocki.net>
Cc: Andreas Noever <andreas.noever@gmail.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>