thunderbolt: Use device links to delay hotplug bridge resume
When resuming from system sleep, the PCI hierarchy is walked down during
the ->resume_noirq phase.  However those subtrees of the hierarchy which
are behind a Thunderbolt PCI tunnel are inaccessible until the tunnel
has been re-established by the NHI driver.  So a Thunderbolt controller's
hotplug bridges (below which the devices behind the tunnel appear) must
delay resuming until the NHI has finished.

That requirement is not enforced by the PM core automatically as it only
guarantees resume ordering between parent and child, and the NHI is not
a parent of the hotplug bridges but a niece.

So far we've open coded this requirement in a PCI quirk, which has the
following disadvantages:

- The code for the NHI resides in drivers/thunderbolt/, whereas the
  quirk for the hotplug bridges resides in drivers/pci/quirks.c, which
  may be surprising and non-obvious in particular for new contributors.

- Whenever support for an additional Thunderbolt controller is added,
  its PCI device ID needs to be amended in both places, which invites
  mistakes.  E.g. commit a42fb351ca1f ("thunderbolt: Allow loading of
  module on recent Apple MacBooks with thunderbolt 2 controller")
  relaxed the subvendor and subdevice ID in the NHI code but forgot to
  also change the hotplug bridge quirk.

- Since the PCI quirk cannot keep any state, it has to search for the
  NHI over and over again on every resume and for every hotplug bridge.

- Thunderbolt PCI device IDs need to be kept in include/linux/pci_ids.h
  rather than drivers/thunderbolt/nhi.h because they're used both by the
  NHI code and hotplug bridge quirk.

Starting with v4.10, the PM core has gained the ability to declare
"device links", i.e. dependencies between devices beyond the mere
parent/child relationship.

Replace the PCI quirk with device links from the hotplug bridges to the
NHI.  The device links are set up when probing the thunderbolt driver
and torn down when it is removed.  The PM core takes care of everything
else.  As a result, the amount of code is reduced significantly.

Signed-off-by: Lukas Wunner <lukas@wunner.de>