PCI: pciehp: Remain in D0 while awaiting command completion
Hotplug controllers may signal an interrupt after a command written to
the Slot Control register completes (PCIe r3.1, sec 6.7.3.2).

If the controller happens to be in D3hot and is PME capable, this
Command Completed interrupt triggers a wakeup to D0 because:

(a) PME is the only valid operation that can be initiated by the
    function when in D3hot (PCI PM r1.2, table 5-4, though Thunderbolt
    controllers do not adhere to this rule).

(b) Hotplug events (including command completion) occurring in D3hot
    cause "generation of a wakeup event (using the PME mechanism)"
    (PCIe r3.1, sec 6.7.3.4).

To be able to write to the Slot Control register, it would be sufficient
to resume the hotplug port's parent to D0 and leave the hotplug port
itself in D3hot.  However it's silly to let the parent go to D3hot
afterwards if we know that it and the hotplug port may be resumed to D0
once the command completes.  Thus, resume the hotplug port to D0 and
hold a runtime PM ref while awaiting command completion.  In the
"nowait" case, schedule runtime suspend after 1 second, which is the
time limit for execution of commands (PCIe r3.1, sec 6.7.3.2).

Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: Mika Westerberg <mika.westerberg@linux.intel.com>
Cc: Erik Veijola <erik.veijola@linux.intel.com>
Cc: Ashok Raj <ashok.raj@intel.com>
Cc: Keith Busch <keith.busch@intel.com>
Cc: Yinghai Lu <yinghai@kernel.org>
Signed-off-by: Lukas Wunner <lukas@wunner.de>