serial: pl011: Deduplicate code to start DMA RX polling
Identical code is present in pl011_dma_startup() and pl011_rx_chars().
By moving this to a separate function, an ugly #ifdef can be avoided in
the upcoming pio_rx() thread function.  Also, the code would have been
indented so far to the right that it wouldn't have fit within 80 chars.

There's an oddity in pl011_dma_startup() wherein the poll timer is
started unconditionally even if pl011_dma_rx_trigger_dma() failed.
I assume this is harmless in that the poll function won't find any
characters in the RX DMA buffer and stop polling after a few tries,
but it nevertheless seems inefficient...  Likewise odd is that polling
is not stopped when RX DMA cannot be retriggered in pl011_dma_rx_irq()
and pl011_dma_rx_callback().

Cc: Chanho Min <chanho.min@lge.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>