ARM: dts: bcm283x: Add overlays for Revolution Pi by KUNBUS
Add a devicetree overlay and a dt-blob.bin source for each member of the
Revolution Pi product family.

The dt-blob.bin is installed on the /boot partition and loaded by the
firmware to set up the initial GPIO pin configuration before powering
up the system.  See here for details:
https://www.raspberrypi.org/documentation/configuration/pin-configuration.md

To determine the specific changes we've made in a dt-blob.bin source,
grep for "RevPi" or compare it to the original provided by the Foundation:
https://github.com/raspberrypi/firmware/blob/master/extra/dt-blob.dts

Briefly, the changes comprise:
* Enable cameras on the RevPi Compact
* Configure UART pins to support a serial early console
* Raise drive strength to achieve higher SPI clock speeds
  (note this setting also affects all other pins in the same bank)
* Set SPI Chip Select inactive on mcp3550-50 (analog input) to prevent
  the chip from going into continuous conversion mode (see datasheet and
  commit c1375d671a35 ("iio: adc: mcp320x: Add support for mcp3550/1/3")

Simon Han contributed portions of the RevPi Connect Flat overlay.

Signed-off-by: Simon Han <z.han@kunbus.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>