PCI: pciehp: Remain in D0 while performing bus reset
When performing a secondary bus reset, two writes to the Slot Control
register are carried out and the port is resumed to D0 for each of them.
In-between, the Bridge Control and Slot Status registers are written.
Keep the port in D0 to ensure accessibility of config space for these
writes.

Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: Mika Westerberg <mika.westerberg@linux.intel.com>
Cc: Erik Veijola <erik.veijola@linux.intel.com>
Cc: Ashok Raj <ashok.raj@intel.com>
Cc: Keith Busch <keith.busch@intel.com>
Cc: Yinghai Lu <yinghai@kernel.org>
Signed-off-by: Lukas Wunner <lukas@wunner.de>