PCI: Export pci_wakeup_bus() & pci_bus_set_current_state()
Commit ea2bf6864e3b made pci_wakeup_bus() & pci_bus_set_current_state()
public for use by drivers which power down an entire bus to D3cold in a
nonstandard way (e.g. not using an ACPI _PR3 object, which would be
handled by the PCI/ACPI core).

We're about to add the first module user of those functions,
thunderbolt.ko, to power down Thunderbolt controllers on Macs,
so export them.

Cc: Bjorn Helgaas <bhelgaas@google.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>