thunderbolt: Obtain PCIe Device/Function number from DROM
Every enabled PCIe Adapter on a Router corresponds to a PCIe Upstream or
Downstream port visible to system software.

Information to associate an Adapter with a PCIe port is contained in the
DROM (USB DROM Specification r1.0 sec 8.5.2).

Obtain the PCIe Device/Function number from DROM and store it in struct
tb_port.

A subsequent commit uses it to discover the struct tb_port corresponding
to a given struct pci_dev.

Signed-off-by: Lukas Wunner <lukas@wunner.de>