thunderbolt: Move upstream_port to struct tb
The code for ICM-controlled tunnel management stores a pointer to the
upstream port in struct icm.  It is going to be needed in the code for
OS-controlled tunnel management as well, so move the pointer to
struct tb which is shared by both tunnel management methods.

Searching for the upstream port previously comprised walking upward from
the NHI until a PCIe upstream port is found.  That seems needlessly
complicated since on all known Thunderbolt controllers, the upstream
port is the NHI's grandparent, i.e. exactly two levels above.  Simplify
the search accordingly.

No functional change intended.

Signed-off-by: Lukas Wunner <lukas@wunner.de>