thunderbolt: Associate PCI devices with PCIe Adapters
Associate PCI devices with corresponding Thunderbolt PCIe Adapters by
storing a pointer to struct pci_dev in struct tb_port.

PCI devices are enumerated asynchronously by the PCI core, independently
from the Thunderbolt driver.  After a tunnel is established, it may take
a while until the PCI devices are visible.  Moreover user space may
remove and rescan PCI devices via sysfs.  For this reason, use a PCI bus
notifier to associate added devices and disassociate removed devices.
On initial switch scan, associate already enumerated devices.

To find the corresponding Thunderbolt PCIe Adapter, the Device/Function
numbers specified in the root switch's DROM need to be available.
On Macs with Thunderbolt 1 that's not the case unless the kernel is
booted by the EFI stub.

Moreover the driver needs to know which tunnels have been established,
which is not the case with ICM-controlled tunnel management, so the bus
notifier is only registered if tunnel management is under OS control.
Perhaps it is possible to retrieve a list of established tunnels from
the ICM firmware, or if all else fails, follow the hop entries in a
PCIe Downstream Adapter's config space to discover established tunnels.
Association would then also work for ICM-controlled tunnel management.

Signed-off-by: Lukas Wunner <lukas@wunner.de>