MAINTAINERS: Add Lukas Wunner as reviewer for thunderbolt
I've been contributing to this driver as well as Thunderbolt-related
bits to the EFI, GPU and PCI subsystems for more than two years, yet
was no longer cc'ed by Intel developers on the changes that went into
4.17.  I literally only learned about their existence from reading the
news.  Add myself as reviewer to make it official that I continue to
care for this driver and would like to be cc'ed on changes.

Cc: Andreas Noever <andreas.noever@gmail.com>
Cc: Michael Jamet <michael.jamet@intel.com>
Cc: Mika Westerberg <mika.westerberg@linux.intel.com>
Cc: Yehezkel Bernat <yehezkel.bernat@intel.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>