PCI: pciehp: Let IRQ handler finish before issuing reset
Before resetting a slot, we disable Presence Detect and Data Link Layer
State interrupts but we make no effort to wait for the IRQ handler to
finish.  There is thus at last a theoretical chance that the handler may
be running concurrently and will react to a reset-induced change in
adapter presence or link state.  That chance grows once the driver is
converted to threaded IRQ handling, so wait for the IRQ handler to
finish after writing the Slot Control register and before issuing reset.

Cc: Rajat Jain <rajatja@google.com>
Cc: Alex Williamson <alex.williamson@redhat.com>
Cc: stable@vger.kernel.org
Signed-off-by: Lukas Wunner <lukas@wunner.de>