PCI: pciehp: Enable/disable exclusively from IRQ thread
Besides the IRQ thread, there are several other places in the driver
which enable or disable the slot:

- pciehp_probe() enables the slot if it's occupied and the pciehp_force
  module parameter is used.

- pciehp_resume() enables or disables the slot after system sleep.

- pciehp_queue_pushbutton_work() enables or disables the slot after the
  5 second delay following an Attention Button press.

- pciehp_sysfs_enable_slot() and pciehp_sysfs_disable_slot() enable or
  disable the slot on sysfs write.

This requires locking and complicates pciehp's state machine.

A simplification can be achieved by enabling and disabling the slot
exclusively from the IRQ thread.

Amend the functions listed above to request slot enable/disablement from
the IRQ thread by either synthesizing a Presence Detect Changed event or,
in the case of a disable user request (via sysfs or an Attention Button
press), submitting a newly introduced force disable request.  The latter
is needed because the slot shall be forced off despite being occupied.
For this force disable request, avoid colliding with Slot Status register
bits by using a bit number greater than 16.

There is a short period on ->probe between addition of the sysfs files
and request of the IRQ when user requests are not yet possible.  An
equivalent period exists on ->remove between freeing of the IRQ and
removal of the sysfs files.  As a courtesy to the user, return -EAGAIN
during those periods by checking the notification_enabled flag instead
of silently ignoring the user's request.

For synchronous execution of requests (on sysfs write), wait for the IRQ
thread to finish and retrieve the result.  There can only ever be one
sysfs write in flight due to the locking in kernfs_fop_write(), hence
there is no risk of returning the result of a different sysfs request to
user space.  Also, synchronize_irq() implies a memory barrier, as does
finishing of the IRQ thread, so there's no need to use something like
xchg() to pass the result from IRQ thread to requester.  If the user
requested slot enablement but the IRQ thread finds the slot unoccupied,
return -ENODEV.

The POWERON_STATE and POWEROFF_STATE is now no longer entered by the
above-listed functions, but solely by the IRQ thread when it begins a
power transition.  Afterwards, it moves on to STATIC_STATE.  The same
applies to canceling of the Attention Button, it likewise becomes an
IRQ thread only operation.

An immediate consequence is that the POWERON_STATE and POWEROFF_STATE is
never observed by the IRQ thread itself, only by functions called in a
different context, such as pciehp_sysfs_enable_slot().  So remove
handling of these states from pciehp_handle_button_press() and
pciehp_handle_link_change() which are called from the IRQ thread.

Signed-off-by: Lukas Wunner <lukas@wunner.de>